Add explicit lookup for all supported mimetypes
